+++ DependencyInjection/OmnipayExtension.php
@@ -11,6 +11,7 @@
 
 namespace ColinODell\OmnipayBundle\DependencyInjection;
 
+use Symfony\Component\Config\Definition\Exception\InvalidConfigurationException;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\Config\FileLocator;
 use Symfony\Component\HttpKernel\DependencyInjection\Extension;
@@ -34,8 +35,41 @@ class OmnipayExtension extends Extension
         $loader = new Loader\YamlFileLoader($container, new FileLocator(__DIR__.'/../Resources/config'));
         $loader->load('services.yml');
 
+        $methods = $config['methods'];
+        $methodNames = array_keys($methods);
+
         // Add configuration to the Omnipay service
         $omnipay = $container->getDefinition('omnipay');
-        $omnipay->addArgument($config['methods']);
+        $omnipay->addArgument($methods);
+
+        if ($disabledGateways = $config['disabled_gateways']) {
+            $omnipay->addMethodCall('setDisabledGateways', [$disabledGateways]);
+        }
+
+        if ($defaultGateway = $config['default_gateway']) {
+            $allowedValues = array_diff($methodNames, $disabledGateways);
+
+            if (!in_array($defaultGateway, $methodNames)) {
+                throw new InvalidConfigurationException(sprintf(
+                    'You cannot specify non-existing gateway (%s) as default. Allowed values: %s',
+                    $defaultGateway,
+                    implode(', ', $allowedValues)
+                ));
+            }
+
+            if (in_array($defaultGateway, $disabledGateways)) {
+                throw new InvalidConfigurationException(sprintf(
+                    'You cannot specify disabled gateway (%s) as default. Allowed values: %s',
+                    $defaultGateway,
+                    implode(', ', $allowedValues)
+                ));
+            }
+
+            $omnipay->addMethodCall('setDefaultGatewayName', [$defaultGateway]);
+        }
+
+        if ($initializeOnRegistration = $config['initialize_gateway_on_registration']) {
+            $omnipay->addMethodCall('initializeOnRegistration', [$initializeOnRegistration]);
+        }
     }
 }