+++ CHANGELOG.md
@@ -4,6 +4,13 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 
 ## [Unreleased]
 
+## [0.3.0]
+### Added
+
+ - Added the ability to define default gateway
+ - Added the ability to disable custom gateways
+ - Optional gateway initialization on registration
+
 ## [0.2.0]
 ### Added
 
@@ -14,5 +21,6 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 
  - Initial release
 
-[Unreleased]: https://github.com/colinodell/omnipay-bundle/compare/0.2.0...HEAD
+[Unreleased]: https://github.com/colinodell/omnipay-bundle/compare/0.3.0...HEAD
+[0.3.0]: https://github.com/colinodell/omnipay-bundle/compare/0.2.0...0.3.0
 [0.2.0]: https://github.com/colinodell/omnipay-bundle/compare/0.1.0...0.2.0